Add coordinator scaffolding (#39)
* Add coordinator scaffolding


* Add github links to TODOs for message types and frost and fire implementations


---------
