+++ sbtc-core/src/signer/coordinator/fire.rs
@@ -1,0 +1,2 @@
+// TODO: FIRE coordination logic
+// https://github.com/Trust-Machines/stacks-sbtc/issues/667
