+++ sbtc-core/src/signer/coordinator/roast.rs
@@ -1,0 +1,2 @@
+// TODO: ROAST coordination logic
+//https://github.com/Trust-Machines/stacks-sbtc/issues/668
